pyproject.toml
[project]
name = "ytube-podcast"
version = "0.1.0"
description = "Youtube to Podcast Generator"
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "piou>=0.20.0",
    "feedparser>=6.0.11",
    "yt-dlp>=2025.6.30",
    "ffmpeg-python>=0.2.0",
    "python-slugify>=8.0.4",
    "python-liquid>=2.0.2",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "BSD-3-Clause"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
podtube = "ytube_podcast.main:run"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
_.env_file = ".env"
podtube = "python src/ytube_podcast/main.py"
